news/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.6
psycopg2-binary==2.9.6
python-dotenv==1.0.0